+++ PlanarAlly/models/campaign.py
@@ -59,7 +59,7 @@ class PlayerRoom(BaseModel):
 class Location(BaseModel):
     room = ForeignKeyField(Room, backref="locations", on_delete="CASCADE")
     name = TextField()
-    unit_size = IntegerField(default=5)
+    unit_size = FloatField(default=5)
     use_grid = BooleanField(default=True)
     full_fow = BooleanField(default=False)
     fow_opacity = FloatField(default=0.3)