+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - GridLayer.size from IntegerField to FloatField
+- Location.unit_size from IntegerField to FloatField
 
 ### Fixed
 - Tokens appear as black/red with all lighting settings disabled [Issues 90/91]